SMT-468 'Something went wrong' looping error when SSO into Mobile Locator App
When using LocatorMobileApp, a client sees the error 'Something went wrong. Please try again'.
The actual error is 'NSURLErrorCancelled'. This happens when one load request is cancelled to handle another load request.
This has been fixed in cordova-ios@4.5.0 by ignoring the NSURLErrorCancelled [https://issues.apache.org/jira/browse/CB-13222].
The LocatorMobileApp uses cordova-plugin-inappbrowser where the same fix was not made. This change ignores the 'NSURLErrorCancelled' and doesn't
handle it as an error.